Move components out of redundant folders
