+++ src/routes/HomePage/HomePage.tsx
@@ -1,5 +1,5 @@
 import { useLoaderData } from "react-router-dom";
-import PostItem from "../../components/PostItem/PostItem";
+import PostItem from "../../components/PostItem";
 import { Post } from "../../types/models";
 
 export default function HomePage() {
@@ -8,10 +8,14 @@ export default function HomePage() {
     // Handle case where posts is not yet loaded or is null
     return null; // or loading indicator or error message
   }
-  posts.sort((a, b) => new Date(b.updatedAt).getTime() - new Date(a.updatedAt).getTime());
+  posts.sort(
+    (a, b) => new Date(b.updatedAt).getTime() - new Date(a.updatedAt).getTime(),
+  );
   return (
     <section className="flex flex-col">
-      <h1 className="mb-10 text-4xl w-full text-center font-semibold pb-8 border-b">Blog Posts</h1>
+      <h1 className="mb-10 w-full border-b pb-8 text-center text-4xl font-semibold">
+        Blog Posts
+      </h1>
       <ul className="flex flex-col gap-4">
         {posts.map((post: Post) => (
           <PostItem key={post._id} post={post} />
